refactor: rename BannerShop and ShopGrid components to PageBanner and ProductGrid for clarity
