+++ src/pages/Cartpage.tsx
@@ -1,10 +1,10 @@
-import { UspBanner, BannerShop } from "../utils";
+import { UspBanner, PageBanner } from "../utils";
 import Cart from "../components/cart/Cart.tsx";
 
 const CartPage = () => {
   return (
     <>
-      <BannerShop title="Cart" />
+      <PageBanner title="Cart" />
       <Cart />
       <UspBanner />
     </>
