+++ src/utils/index.ts
@@ -6,15 +6,14 @@ export { default as ProductCard } from "../components/home/ProductCard.tsx";
 export { default as Img } from "../components/common/Image.tsx";
 export { default as Heading } from "../components/common/Headingtxt.tsx";
 export { default as Input } from "../components/common/Input.tsx";
-export { default as ShopGrid } from "../components/home/Shopgrid.tsx";
+export { default as ProductGrid } from "../components/home/ProductGrid.tsx";
 export { default as Button } from "../components/common/Button.tsx";
 export { default as LoadingSkeleton } from "../components/common/LoadingSkeleton.tsx";
 export { default as UspBanner } from "../components/common/UspBanner.tsx";
-export { default as BannerShop } from "../components/common/BannerShop.tsx";
+export { default as PageBanner } from "../components/common/PageBanner.tsx";
 export { default as FilterSortBar } from "../components/filters/FilterSortBar.tsx";
 export { default as Pagination } from "../components/common/Pagination.tsx";
 export { default as Breadcrumb } from "../components/common/Breadcrumb.tsx";
 export { default as Icons } from "./Icons.tsx";
-export { default as Related } from "../components/detailpage/RelatedProducts.tsx";
+export { default as Related } from "../components/detailpage/ProductRelated.tsx";
 export * from "./helperFunctions.ts";
-
